started implementing simulation for eye movements
The current concern is that effects in primary motor cortex are due to actually
executed eye movements towards the shown dots. If you take this thought
seriously, it means that the visual input should change after every eye
movement. Specifically, the motor commands necessary for moving the eyes to the
dot position should change, because starting position of the eyes is not at the
centre of the screen anymore. In some cases it may then be that the movement
to a dot on the left side may actually be done by commanding the eyes to move
right (if the previous dot was further to the left than the current). This is
made pretty complicated once you consider the delay between dot occurrence and
movement onset which is assumed to be 300 ms or 400 ms in our case.

This is the first working implementation of a simulator for delayed eye
movements towards the presented dots.