under the assumptions I made, eye movement motor commands are move_x/y
This was my intuition all along, but I wanted to get to it through a verified
path. My previous simulation of eye movements only simulated the eye movements
themselves, i.e., how the eye position changes throughout the trial assuming
that it follows the dot positions with some delay. If you additionally assume
that eye movements are planned with a reliable, current estimate of the eye
position, a reliable estimate of dot positions in screen coordinates and eye
movements are faster than dot position changes, the motor commands to move the
eye from the current position to the desired dot position are simply the move_x
and move_y values that I previously computed, i.e., the vectors from the
previous to the current dot position. That these are delayed by 300 or whatever
ms doesn't matter as long as they are delayed for each dot equally.

The only difference between move_x and the desired saccade motor command is
that move_x is 0 for the first dot while it should be the x-position of the
first dot. To account for that I added new 'saccade_x' and 'saccade_y'
regressors.

It turns out that saccade_x and dot_x have a pretty high correlation around
0.76 which should not have surprised me as I had previously found that move_x
is quite strongly correlated to dot_x. It remains to be seen whether it is
still possible to decide whether dot_x or saccade_x better describe the data.