investigating correlations with saccade_x in primary motor cortex
The argument is that, if the signal in primary motor cortex reflects mostly eye
movements, then it should also more strongly correlate with the eye movement
adjusted saccade_x regressor than with dot_x itself. To check whether this is
true, I produced a plot in which I compare the GP-estimates of the underlying
relations at 400 ms after dot onset. This clearly indicates that dot_x is more
strongly associated with the signal in primary motor cortex than saccade_x. I
also computed statistics for the null hypothesis that the corresponding linear
correlations are equal. Both, confidence intervals and p-values, indicate
strong evidence for a difference in correlations in favour of dot_x.