comparing accumulated (sum_dot_x) and momentary (dot_x) evidence
For a status report to Stefan I decided that it would be helpful to collect
arguments for a representation of accumulated evidence in primary motor cortex,
because that would indicate that the motor cortex not only reacts (by e.g.
preparing eye movements) to the change in dot position. So I extended the code
used to compare saccade_x and dot_x regressors.

Different from my previous comparisons I use GPs now to check the underlying
mean relationship between the regressors and the signal. This, surprisingly,
shows that in the range -100 to 100 px the signal behaves almost equal for both
dot_x and sum_dot_x. That is the case even though the correlation between dot_x
and sum_dot_x is smaller than that between dot_x and saccade_x (0.67 vs. 0.77).
dot_x and sum_dot_x differ only at the edges. Specifically the signal rises for
sum_dot_x values above 200 (for the sign corresponding to the contralateral
side) while for dot_x above 180 the signal appears to drop. This suggests that
the accumulated evidence explains slightly more of the signal. This difference,
however, is based on the large values for which there are no actual
x-coordinate values, because 99% of these are below 185 px and their maximum is
at 232. To see the edges I also had to increase the quantile range of regressor
values to include in the analysis.

Note also that sum_dot_x has heavier tails than dot_x. This means that when you
normalise the regressor values (by deviding through standard deviation), you
strongly squish the centre of the distribution. When you plot this, you then
get in the right hemisphere an s-shaped curve for sum_dot_x in [-2.5, 2.5]
while dot_x shows a very linear shape in the same range indicating that perhaps
dot_x has the clearer relation to the signal. However, knowing the behaviour of
the signal for large values of dot_x (approximately), I find this misleading,
because a normalised value of 2.5 only correponds to a dot_x value of ca.
191 px which excludes the large x-values.

I also tested the results for when I exclude data points for which sum_dot_x
and dot_x have similar values, i.e., I only include data for which there is a
large difference between the regressors. I set the threshold for the difference
so that 80% of all data points are excluded. This tends to make it clearer that
sum_dot_x explains a larger part of the signal, especially in the right
hemisphere, but I'm unsure whether through this selection of data/trials I
select a specific subset of data with interacting other properties, e.g., do I
specifically exclude negative dot_x values, but not sum_dot_x values?

I wonder whether the value of sum_dot_x for which the signal reaches a maximum
corresponds to the decision threshold? Perhaps I can estimate this on a per
subject basis and then compare it to the decision threshold estimated from the
model. Note, however, that in the left hemisphere I don't see a clear maximum
as the signal still appears to rise for the highest plotted regressor values.
How would I treat this case? Well, if I simply set the xrange larger to 0.001
and 0.999 quantiles, a peak becomes visible, but it also becomes clear that
the sparse GP appears to fall back to predicting 0 for out of range values (for
dot_x) and I wonder whether the peaks simply occur, because  there is a pull
from the prior towards 0 and less data is available for large regressor
values (some of these values occur only once per subject).