shape of single-subject correlations between dot_x and signal in M1
In commit e8b142d1 I had speculated about whether the peaks of the sum_dot_x
correlations in primary motor cortex could reflect the decision threshold and
whether I could see a correlation between the peaks and the inferred decision
thresholds from the model across subjects. I had a look at the shape of
correlations for individual subjects up to subject 17. There is huge variance
over the shape of fitted sparse GPs. Especially, it won't be possible to
reliably estimate the peaks I thought about for individual subjects.

One complication is that in individual subjects the normalised mean currents
appear to be strongly offset. E.g. for subject 17 the right hemisphere currents
appear to have a mean level around 0.1 while the left hemisphere currents are
close to 0. I should be able to fix this by introducing a corresponding offset
into the GPs, but this only solves a small part of the causes for the variable
shape of the function between x-coordinates and signal.