checking results with different GPs (Matern, bias kernels)
bias
----
As I noticed that for some subjects the normalised mean currents may have
offsets, I decided to include the possibility of such an offset into the GPs. I
did this using the bias kernel which adds a constant to the covariance function
where this constant is the variance of a 0-mean Gaussian prior over the
intercept. I had the impression that including the bias kernel improved
estimates for individual subjects, but I did not quantify this in any way and
couldn't name concrete examples.

For the pooled data, results are very similar with and without bias kernel
except that the uncertainty bands may be a bit tighter with bias kernel,
especially for sum_dot_x.

Matern
------
In the GP book Williams and Rasmussen write that the squared exponential
covariance function has been criticised for being too smooth for 'real' data
and that the Matern-class covariance functions may be more suitable. I compared
the previous results with one using the Matern32 covariance function. As
expected, the overall shape is very similar, but there are a few more (smooth)
local bumps in the Matern results. The bands of uncertainty overlap a bit less
in the Matern results which appears to be mostly because they are a bit tighter
for dot_x for large x-values.

constant functions
------------------
Going through GP fits of single subjects (rbf with bias) I noticed that for
some subjects the predicted function is just a constant which may be close to
0. In some cases the bands of uncertainty were also surprisingly tight
(e.g., <0.001 for subject 27, left hemisphere). I checked whether something is
going wrong with this data, but found that everything works fine. It simply is
the case that the marginal likelihood appears to be higher for GPs which set
the bias variance to something very small so that the GP offset is bound to 0
and at the same time set the rbf variance very small. What's a bit strange is
that this can occur together with a non-zero estimate for the linear regression
coefficient. Why does the GP estimate a linear function with slope=0 while the
maximum likelihood slope of a linear regression is estimated to be different
from 0? I guess the marginalisation within the GP has something to do with it.

recovery tests for a constant function
--------------------------------------
Out of curiosity I tested whether the GP would recover a constant function
under noise conditions as in the data. To do this I fitted the GP as such:

gpm = fitgp(np.tile(np.linspace(0, 1, 2000), 2),
            np.random.randn(4000) * 1 + 0.05,
            np.linspace(0, 1, 15))

In the 4 tests I ran the GP always recovered a constant function with bands of
uncertainty including the true value of 0.05, but the mean was at 0.055, 0.068,
0.064 and 0.025.

For a previous setup with noise std of 10 and mean 0.34 I also observed
a non-constant function in one repetition in which the mean ranged from -0.12
to 0.5, but also there the mean 0.34 was included in the bands of uncertainty
for all x-values as far as I could tell from the plot. This means that even
with 4000 data points the GP does not always recognise that the underlying
function is constant when the constant is below 5% of the noise std, but
otherwise the result was still reasonable.