make GP functions compatible with python2 environment + allow mlog10p in fdr
Only import GPy and define the GP functions, when GPy exists in environment.
Also, there were some issues with numpy in the mne2 environment so I added the
number of elements in linspace as keyword-argument, but I don't remember
whether that was necessary in the end.

You can now also use mlog10p (instead of mlog10p_fdr) as measure in
get_fdrcorr_clusters.