use standard inference and plots for early visual effects
I did this analysis originally with the VEF-files, but for the DySCo
presentation I wanted to get plots that are consistent with the others.
Specifically, I now apply FDR-correction as for the main analyses. To reuse
most code I also ran standard sequential inference for the V1/V2 vertices to
get the second-level results. Note that I only used data up to 400 ms into the
trial and dots 1-3 resulting in the first 200 ms after dot onset. Results are
in:
	source_sequential_201711271306.h5

Because I could not plot both hemispheres in one figure (V1 cannot be seen
properly), I decided to produce two figure windows, one for each hemisphere, to
avoid problems with timing in TraitsUI as warned about in pySurfer. Probably I
should do the same in show_motor_vertices, or at least also introduce a loop
over hemispheres.