plot dot_y and response effects in the brain + bugfix
For my DySCo presentation I generated new figures, especially for
response-related effects. I now explicitly allow the colormap to be generated
from a different regressor than the chosen one to unify colormaps across
regressors.

As for dot_y only POS1 becomes significant at 320 or 330 ms, respectively, I
now also allow that non-significant effects are shown, but with the same
colormap. For absmean as show_measure this shows a few more effects also at 170
ms. These must be non-significant, because larger across-subject variance than
for POS1 around 320 ms.

For the first dot-aligned response-effects no significant results exist after
FDR. Hence, I choose the colormap based on the quantiles of the selected
measure. The particular quantiles are chosen to exclude most spurious results
as per my intuition.

I also made a new analysis for response-aligned response effects for the talk.
Inference results for that are in

	source_singledot_201711281123.h5 (response-aligned in [-200, 50])

bugfix
------
I forgot that the file names generated by save_image_sequence will
contain the index into the time vector and not directly the time. Previously,
this was not an issue, because the times I looked at started at 0 so that 120
ms corresponded also to index 12, but for response-aligned times index 0 may
correspond to time -200. Now I account for that when stitching the generated
files together.