repeat model inference for behavioural data because of wrong veps
Checking something for Shakoor I revisited my setting of veps for the reaction
time models. It turned out that I misinterpreted veps slightly leading to a
factor of 2 too much in its definition. The issue was that veps is the
normalising constant of the uniform distribution over the space in which
samples are accepted, but I included the RT space of the unchosen option in
which no samples are accepted.

I reran the inference with the corrected veps. Results are in
	infres_basic_201712011831.npz
	infres_collapse_201712012202.npz

Note that the change in veps only increases marginal likelihood estimates by a
constant depending only on the number of data points. So I could have
equivalently added the constant to the old results. Also, model comparisons
should be unaffected as the constant is added equally to all marginal
likelihoods. Parameter estimates are unaffected by veps anyway.