make cleaner version of accumulated evidence over time regressors (sumx_time)
In preparation of my response-aligned analysis of accumulated evidence I
cleaned up the accev_time regressor with the following changes:

sum_dot_x instead of sum_dot_x_prev
-----------------------------------
It didn't make much sense to use sum_dot_x_prev in a time-based analysis: It's
just a time-shifted version of sum_dot_x with a value missing for the first
dot. As I anyway have the delay to shift values in time, there really was no
point.

proper indexing
---------------
It turns out that my way of finding the dots corrresponding to a particular
time point had some numerical issues: Real numbers may not represent a value
such as 0.3 precisely as 0.3, but perhaps as 0.2999999373. Using floor division
on such values does not always give the expected result. I was only able to
circumvent this by transforming the involved values to large enough integers.

unobserved values and caching
-----------------------------
I now allow to specify an unobserved value, i.e., the value that should be used
at time points for which no regressor value is available. This occurs for time
points before the specified delay and after response+delay, because the dot
vanished at the button press. Previously I set the value for the regressor
after response.delay to the value at response+delay, but I now thought that
indicating that there is no value at this time point is better. As a
side-effect the accev_time regressors now compute very fast so that I don't
need the caching of pre-computed regressor values. I suspect that this is
because I don't look up the values at RT+delay anymore, but it could also be
due to some improvements in pandas after I updated it. Anyway, I didn't check.