analysis with new time regressors, first dot onset aligned, x-flip, bugfix
bugfix: didn't actually drop nans in params before second-level analysis

generalisation to time-based design matrix
------------------------------------------
I now do analysis including 4 time-based regressors (dot_x, dot_y, and the new
perceptual update regressors). Any combination of time-based regressors can now
be chosen for the analysis, as conveniently returned by
subject_DM.get_trial_time_DM.

choose between response- and first-dot-onset-aligned
----------------------------------------------------
As the difference is only in indexing only few changes in the code had to be
made to allow for that.

x-flipping distinguishes motor response from evidence representation
--------------------------------------------------------------------
There is a very strong linear relationship between dot-x and the M1-signal 30
ms after the response, i.e., when the correlation between dot-x and M1-signal
is largest. The linear relationship is deceiving, though, as it apparently only
results through the probabilistic association between dot-x and the response.
This can be seen, when you sign-flip all those dot-x that have a sign
contradicting the response. At 30 ms these are only 4719 of 16242 values. So
the sign-flipping shouldn't destroy the relation between x and M1 completely,
but it brings x in alignment with the response. The result clearly shows that
the peak at 30 ms is not driven by a linear relationship between x and M1, but
rather by two distinct activity levels for the two responses. For data before
the peak this is much less clear and I'm considering implementing a
probabilitic model which could differentiate a linear relationship from the
more step-like.

Results
-------
are described in more detail in my email to Stefan and the associated document
(response-aligned_evidence...). As expected the dot_x time course in M1 is very
similar to that of sum_dot_x. In the GP-plots the other regressors tend to stay
below a magnitude of 0.02 which appears to be a level over which results become
reliable, although still weak.

For v23ab I see two outstanding peaks for dot_x in the response-aligned
analysis: one just after the response as in M1 and one ~ -670 ms before the
response. I investigated whether the -670 ms peak could be explained by a
single dot effect, but didn't find anything.

When aligned to first dot onset, the results in M1 show that the dot-x-effects
are strongest early in the trial (starting at 400 ms for the same delay) and
then drop to 0. I thought that this also is a good indicator that M1 actually
tracks evidence and does not only prepare the response. The reason the signal
drops could be that M1 becomes uninterested in the signal as soon as the
response has happened, i.e., it doesn't track evidence after the response and
because more and more responses have happened as time passes, the correlation
with x drops. I also thought that it is a good sign that I don't see a shadow
of the response peak in this time course, i.e., the RT mode is at ~ 900 ms so,
if this motor effect had a strong influence on the x-correlations, I should
observe at least a small bump around 900 ms, but I don't.