exclude late trials, check sum_dot_x, fix trial_counts, add area label
area label: is now in title for when I compare several figure windows
trial_counts: now directly counted from data for GLM per subject

exclude late trials
-------------------
To test my hypothesis that the effects in M1 drop from first dot onset because
in more and more trials responses have happened I simply excluded trials with a
certain minimum distance to the response. For that I included the new option
'toolate' giving the response-aligned time from which trials are excluded.

When I exclude all trials with response-aligned times > -100 ms, the gradual
drop of effect magnitudes through first-dot-onset-aligned time does not
disappear! It seems that the early strong effects become stronger and for late
times the effects do not drop to 0, but stay on a weak level.

check sum_dot_x
---------------
I wondered whether I see the same pattern for sum-x instead of x (I started
with a response-aligned analysis of sum-x, but hadn't looked at
first-dot-aligned sum-x, yet). I was surprised to see that the effect of strong
initial magnitude with drop towards 900 ms is considerably stronger for sum-x!
Indeed mean beta-values reach up to 0.1 for sum-x at 0 ms which is 2/3 of the
beta values reached for the motor response peak (~ 0.1/0.15) whereas this is
only 1/3 for the x-coordinate (~ 0.04/0.12).

I initially thought that this was a great finding supporting the idea that M1
represents accumulated, not momentary evidence, but there is an interaction
with the scale of regressor values that I didn't consider: I normalise
regressor values across all trials and times, but sum-x values tend to be
smaller at early time points and larger at late time points. So the GLM finds
larger beta values for sum-x at early time points. This doesn't mean that sum-x
explains more of the signal at early time points, but it explains why these
seemingly large early effects were not significant at many time points. I need
to find a better measure to judge these effects.