local normalisation, online garbage collection, restructuring, file renaming
Normalisation and interpretation of betas
-----------------------------------------
Following the potentially misleading results for sum-x I decided to implement
local normalisation: Instead of normalising the design matrix across all
possible regressor values I now normalise it locally for each GLM (at
individual time points within subjects) and do the same (at least the scale)
for the data. The resulting interpretation of the regressor betas is: For a
standard unit step in the regressor at that time point you will make a
beta * data-std-sized step in the data at that time point, i.e., you can
interpret beta as a measure for the proportion of variability in the data
explained by the regressor.

In previous analyses with 'global' normalisation this interpretation was
confounded by the varying scale of the regressors across time points as well as
a, to a much lesser extent, varying scale of the data across time points: The
same interpretation in principle applied, but the betas were only estimated
from local subsets of the regressors and data available to the least squares
optimiser of the GLM for a given time point. The high betas for sum-x at early
time points with global normalisation suggested that there is a very strong
representation of sum-x in the data, but this interpretation is not warranted,
because it generalises a very restricted view of small sum-x values to the
whole range of them. Global normalisation may be ok in the sequential analysis
in which I stack several time points including high sum-x values, but here it's
misleading, and even in the sequential analysis local normalisation may be
preferred for a secure interpretation of the results (but results should be
similar for local and global normalisation).

As could be expected, the difference between local and global normalisation is
not so big for the non-accumulating regressors, as their scale does not vary
strongly across time points, but local normalisation makes results for
sum_dot_x very similar to those for dot_x.

other changes
-------------
When running the script twice in parallel I ran into memory issues. Including
an explicit call to garbage collection solved that.

- made file names more informative.
- removed old code for checking GPs
- moved plot of single time point to end so that it doesn't break execution of
  the script when, e.g., a time doesn't exist in the specific analysis