testing model for estimating linearity of a data set with Stan
This is preparation for checking whether the signals I observe are actual
correlations, i.e., linear relationships between regressors, or are better
described by a step-function. I have first implemented a single-level model in
Stan and verified that it returns low values of l(inearity) for step-like data
and high values for linear data (l<0.1 vs. l>1, respectively). l is simply the
inverse slope, or lengthscale of a sigmoid function where values of l towards
0 make the sigmoid increasingly step-like. Assumption is that x-values are
normalised to have std=1. Additional parameter 'beta' gives the height of the
level that is reached in y-coordinates.

practical issues with Stan
--------------------------
For some reason the models do not compile (compilation breaks with
uninformative error), when I invoke 'matplotlib qt5' before compilation in the
ipython interpreter. I have asked about this on the pystan issue tracker on
github.

implementing positivity constraints in hierarchical model
---------------------------------------------------------
In the hierarchical model I have l = mu + x * std where x can be negative,
but l should be positive. This is a recurring issue that I came across already
in DDM modelling. I have now found an easy way to resolve this following this
thread:

https://groups.google.com/forum/#!topic/stan-users/JI6yrSZz_Xg

The trick is to use lower-bound constraints on l which depend on mu and std, as
demonstrated in the corresponding model code.

An initial test with the hierarchical model showed that inference is pretty
slow for number of subjects and trials similar to the actual data, but the
results are pretty ok. There are some sampling issues with parameters for some
subjects (Rhat > 1.1). These most likely result from data that is uninformative
either about beta or the intercept, but it's a bit strange that issues occur in
either the one or the other. Anyway, I realised that I do not want to allow
linearity to differ across subjects: either the effect is linear, or step-like,
but not one for one subject one for another subject. So I will take the
subject-level linearities out and hope that it also helps to resolve some of
the issues with sampling.