first run of full inference of linearity for primary motor cortex
After testing the hierarchical step model on simulated data, I have now
implemented full inference with the model across time points of an area. The
inference will be done with and without sign flips aligning the x-coordinates
with the choice so that I can compare the two. As I only have the time course
for a single area, I can store the full set of samples from the posterior
(excluding warm-up). The code closely follows that of
infer_respaligned_source.py.

It takes roughly 21 hours to run through 111 time points [-700, 400] of one
source label.

Results
-------
can be found in
	source_linearity_201712181904.h5

So far only times in [-700, 340] of the left hemisphere finished computing. The
results confirm some of my suspicions:
- from -30 to 170 ms after the response linearity l is almost exactly equal to
  0 for the flipped data indicating that M1 onyl represents the response then
- there is a spike in linearity around 230 ms, but a drop back close to 0
  afterwards for flipped data indicating that the after-response overshoot also
  just reflects the response itself
- at these time points the magnitude of the estimated betas is equal or larger
  for the flipped x-values indicating that the flipped dot-x explains more of
  the signal than dot-x itself, but this effect is very small
- from -700 to about -190 ms linearity tends to be high between values of 1 and
  1.5 (except for a drop to 0.15 around -300 ms) indicating that at these time
  points a gradual relationship between dot-x and the M1-signal best explains
  the data
- this is further supported by the tendency of beta-magnitudes to be higher for
  the original compared to the flipped x-values