problems with sigmoid-model + testing new linear-step-mixture-model
The overall results for the sigmoid-based linearity model made quite a lot of
sense to me: the estimated beta roughly showed the effects that I had seen for
the linear model before and the estimated linearity indicated that only close
to the response and after it the signal is step-like. However, then I noticed
that I couldn't really say that the signal was linear for the early time points
despite a high inferred linearity: For very small effects the model has large
uncertainty over linearity which makes it look like you are seeing a linear
effect. The other side of the same coin is that also for clearly linear signals
the model tends to still have a relatively uncertain linearity estimate as
always large values of linearity are plausible in combination with large
values of beta.

New linear-step-mixture model
-----------------------------
I always had in mind that the alternative to my continuous sigmoid approach was
to simply compare the linear vs. a step model, but I didn't want to run two
models separately and like the idea of having a continuous estimate of
linearity. So I figured a way to combine these potentially contradicting
approaches: model testing through a mixture model. This has been advocated by
Christian Robert and others in an oldish arXiv paper from 2014 (Kamary et al.,
1412.2044). You'll simply estimate the mixture probability lambda for two
likelihoods, one implementing the linear model, the other implementing the
step model. The paper suggests that this is a good alternative to model
comparison and you get lambda as a continuous estimate of linearity of the
signal.

It was pretty straightforward to implement this mixture model in Stan.
Inference takes longer than with the sigmoid model (about 2 times?). In tests
with simulated data I had the impression that the inference works reasonably:
A flat version of the model seemed to correctly identify a linear and a step
data set through lambda. The beta-parameter of the winning model was well
estimated close to its true value while the beta of the other model had very
large uncertainty which I found plausible as it shouldn't be well defined given
that it's not used to explain the data.

In a hierarchical version of the mixture model I tested inference for data that
does not hold any relation between x and y, i.e., with beta=0. I observed
several different behaviours. It could happen that one of the two included
models wins in all chains. Then lambda looked like it had convincingly
identified one of the models to be true for the data. At least the estimated
beta for the winning model then quite precisely indicated a null-effect (0
within 95% PPI, e.g., [-0.04, 0.02]). The other case was that in at least one
chain the other model wins with a correspondingly more spread out distribution
of lambda. Because the variance of the beta of the loosing model is high, the
betas of both models will have large variance when the chains are mixed. These
outcomes are not ideal. I would much prefer a within-chain mixing between the
two models, but both conditions appeared to be suitably distinct to identify
irrelevant effects. So I would have accepted that behaviour as a tolerable
shortcoming.

Finally I tested the behaviour of the hierarchical mixture model on two real
data sets: 1) -700 ms in left M1 and 2) 0 ms in left M1. My previous analyses
indicated for 1) a small linear effect and for 2) a sizable linear effect for
the original dot-x and a sizable step effect for the flipped dot-x. To my
surprise, the mixture model identified a small step effect for 1) in two
independent repetitions of inference. In a 3rd repetition only one of the 4
chains indicated a step effect while the others indicated a linear effect.
Demonstrating the unreliability of the results. For 2) and the presumably large
effects inference did not work better: In one run with the original dot-x half
of the chains indicated a linear effect while the other half indicated a step
effect, although for the linear effect chains the lambda at least appeared to
be estimated more decisively close to 1 than it was estimated close to 0 in the
step chains. In another run with the flipped dot-x the chains were again evenly
split between linear and step effects.

Taken together these tests demonstrated that Stan's MCMC algorithm is not able
to properly mix between the two models. Depending on the random initialisation
the sampler will remain on one model despite there being reasonable evidence
for the other model. The result is garbage. So I'm still lacking a model that
with certainty indicates linearity for linear data, with certainty indicates a
step-relation for step data and is appropriately uncertain for weak effects.

side-note: ab_beta
..................
Kamary et al. suggest a Beta-prior on lambda with alpha=beta=0.5. This strongly
favours lambda values close to 0 and 1 which probably is supposed to ensure a
more binary outcome implementing the assumption that only one of the models can
be true. Given the within-chain mixing problems, my intuition said that this
could be a problem for the sampling. Hence I made the prior flatter with
alpha=beta=0.8 for the tests on the real data. As reported, this did not help
either. In a final test I set ab_beta=0.95 on the 0 ms data with flipped dot-x.
The result still had two chains for each model with no samples crossing over to
the other model within a chain (crossing here would mean that lambda comes
closer than 0.1 to the opposite limit).