linear vs. step models with loo-cross-validation in M1
Following the successful test of the cross-validation based model comparison
for linear and step models I have now run the corresponding analysis on the
signal in primary motor cortex. The Stan sampling for one model and data set
runs for about 5 mins so that the whole analysis (2 models + 2 choice flips +
111 time points) runs for about 2 days for one brain area. The results follow
the expected pattern, but also disappoint a bit, because the model comparison
indicates no difference between linear and step models for early time points
before the response. This is because the effects are so small and noisy that it
is not possible to reliably say whether the signal rather follows a linear or
step-like relationship with the x-coordinate. Nevertheless, the inferred
beta-values may still indicate that the relation is linear earlier than -200 ms
before the response, because they then are larger for the original x than for
the choice-flipped x and this is reversed for values around the response. I
show this statistically in a new figure which plots the distribution across
time points of the difference between posterior mean betas. Some more details
follow.

Inference
---------
I have adapted the inference code to the new model comparison approach. I
decided to only store the samples for the population-level effect beta, but
also store Rhat and neff diagnostics for it. This is because I can only extract
samples from a single parameter, when the samples are first permuted across the
chains before they are returned to me. This means that I cannot store the
actual sequence of samples sorted by chains. I did not want to extract all
samples from all parameters, which would prevent permutation, because that took
much longer.

I now normalise also after taking the sign of x in the step model. This is
supposed to ensure that beta_lin and beta_step are more comparable (reflecting
estimates of the correlation). I had the impression that their values are
indeed more similar now, but did not investigate any further.

For my analysis I now use the mean of the leave-one-out cross-validation values
instead of the sum, because the number of data points varies across time
points. I have adapted the computation of the standard errors accordingly.
Everything appears to work fine.

Figures
-------
I updated the main figure to the new results format and deleted the remaining
ones as they were not necessary anymore.

The new figure investigates whether the beta estimates are larger for the
original or the choice-flipped x. I decided to do this across time points, but
compare within time points as this best reflects what you can see in the main
figure. So for each time point the two means of the beta samples enter the
analysis. I can then apply a paired-sample t-test across time points and plot
the corresponding differences in the figure.

Results
-------
are in
	source_linearity_201801041820.h5
	(dot_x_time, delay 400, area 4, [-700, 400])

I decided to only include time points up to -700 ms before the response to
reduce computation time. As mentioned, the model comparison cannot identify the
correct model for early time points. More information can be found in my
report: response-aligned_evidence_Email_to_Stefan_20180108.pdf