new singledot analysis with long epochs, RT-based exclusion and new atlas
This is the first of the reruns of the analysis for the long epochs that I now
got. Additionally I have now implemented that trials are excluded whenever they
get too close to the response (here toolate=-200 based on the response-aligned
analysis suggesting that response signal starts around 200 ms before the
response). Further, source analysis is now based on the newer HCPMMP1_5_8
atlas. Because of the longer epochs I have included the first 7, not only the
first 5 dots as regressors into the analysis.

Results are in:
	source_singledot_201801171913.h5

and needed approximately 2 days per permutation to compute, i.e., roughly 8
days in total.

I am a bit surprised how quickly the second-level beta-estimates become
unreliable when progressing into the trial: It looks as if already from about
600 ms the variance of the betas in the permutation appears to be increased
indicating that beta estimates might become unreliable. This is although the
mean number of trials per subject is still 385 at 600 ms. From about 1000 ms
the beta estimates appear to be garbage (mean number of trials at 1000 ms is
182).

The removal of trials close to the response consequently diminishes the effect
of dot-x considerably, especially for dot 4 onwards (in left M1). I believe,
however, from the figures I've seen that the effect is still clearly present,
although small.